requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.1
django-redis==4.9.0
djangorestframework==3.8.2
drf-extensions==0.4.0
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-tasks==0.4.0
googleapis-common-protos==1.5.8
grpc-google-iam-v1==0.11.4
grpcio==1.19.0
idna==2.8
protobuf==3.7.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
redis==2.10.6
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
